docs: api events
